Add conformance section
Fixes "Normative references in informative sections are not allowed" warning
messages reported by Redpec.

The warnings appeared because Respec assumes that a spec is informative-only
in the absence of a conformance section, see discussion in:
https://github.com/w3c/respec/issues/2580

Also fixes the References section as some of the specs were incorrectly listed
as informative for the same reason.